Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15556 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/nexusSrc/napi.c
@@ -829,7 +829,7 @@ static int analyzeNapimount(char *napiMount, char *extFile, int extFileLen,
 char *nxitrim(char *str)
 {
       char *ibuf = str;
-      size_t i = 0;
+      int i = 0;
 
       /*
       **  Trap NULL
@@ -848,7 +848,7 @@ char *nxitrim(char *str)
             /*
             **  Remove trailing spaces (from RMTRAIL.C)
             */
-	    i = strlen(str);
+	    i = (int)strlen(str);
             while (--i >= 0)
             {
                   if (!isspace(str[i]))
